refactor: move similarity and greatest sql functions into dedicated file
