+++ src/main/kotlin/se/uulm/snowballr/backend/repository/RepoHelper.kt
@@ -1,9 +1,6 @@
 package se.uulm.snowballr.backend.repository
 
 import org.jetbrains.exposed.v1.core.ColumnSet
-import org.jetbrains.exposed.v1.core.CustomFunction
-import org.jetbrains.exposed.v1.core.Expression
-import org.jetbrains.exposed.v1.core.FloatColumnType
 import org.jetbrains.exposed.v1.core.JoinType
 import org.jetbrains.exposed.v1.core.Op
 import org.jetbrains.exposed.v1.core.ResultRow
@@ -171,21 +168,6 @@ fun List<ResultRow>.toProjectPaperWithExternalIds() = first().toProjectPaperWith
 private fun List<ResultRow>.toExternalIds(): List<ExternalId> =
     filter { it.getOrNull(PaperHasExternalIdTable.type) != null }
         .map(ResultRow::toExternalId)
-
-/**
- * Trigram similarity between [left] and [right] as provided by the `pg_trgm` extension.
- *
- * The result ranges from 0 (nothing in common) to 1 (identical).
- */
-fun similarity(left: Expression<String>, right: Expression<String>): CustomFunction<Float> =
-    CustomFunction("similarity", FloatColumnType(), left, right)
-
-/**
- * Largest value of the passed [expressions].
- */
-@Suppress("SpreadOperator")
-fun greatest(vararg expressions: Expression<Float>): CustomFunction<Float> =
-    CustomFunction("GREATEST", FloatColumnType(), *expressions)
 
 fun ColumnSet.joinPaperHasExternalId() = this
     .join(
